+++ demo.php
@@ -29,10 +29,6 @@ class Container
     }
 }
 
-if (!isset($_ENV[Environment::ENV_VAR_NAME])) {
-    $_ENV[Environment::ENV_VAR_NAME] = Environment::QA;
-}
-
 $apiKey = isset($_ENV['ROLLOUT_API_KEY'])
     ? $_ENV['ROLLOUT_API_KEY']
     : DEFAULT_API_KEY;
@@ -42,7 +38,15 @@ $devModeKey = isset($_ENV['ROLLOUT_DEV_MODE_KEY'])
     : DEFAULT_DEV_MODE_KEY;
 
 $roxOptionsBuilder = (new RoxOptionsBuilder())
-    ->setDevModeKey($devModeKey);
+    ->setDevModeKey($devModeKey)
+    ->setNetworkConfigurationsOptions(new NetworkConfigurationsOptions(
+        'https://api.test.rollout.io/device/get_configuration',
+        'https://rox-conf.test.rollout.io',
+        'https://api.test.rollout.io/device/update_state_store',
+        'https://rox-state.test.rollout.io',
+        'https://analytic.test.rollout.io',
+        'https://notify.bugsnag.com'
+    ));
 
 echo '<pre>';
