+++ src/Rox/Core/Network/ConfigurationFetcher.php
@@ -11,43 +11,9 @@ class ConfigurationFetcher extends ConfigurationFetcherBase
      * @param array $properties
      * @return string
      */
-    private function _getPath(array $properties)
-    {
-        return $properties[PropertyType::getAppKey()->getName()] . '/' . $properties[PropertyType::getBuid()->getName()];
-    }
-
-    /**
-     * @param array $properties
-     * @return string
-     */
     private function _getCDNUrl(array $properties)
     {
         return $this->_environment->getConfigCDNPath() . '/' . $properties[PropertyType::getCacheMissRelativeUrl()->getName()];
-    }
-
-    /**
-     * @param array $properties
-     * @return string
-     */
-    private function _getAPIUrl(array $properties)
-    {
-        return $this->_environment->getConfigAPIPath() . '/' . $properties[PropertyType::getCacheMissRelativeUrl()->getName()];
-    }
-
-    /**
-     * @return array
-     */
-    private function _preparePropsFromDeviceProps()
-    {
-        $queryParams = $this->_deviceProperties->getAllProperties();
-        $queryStringParts = $this->_buid->getQueryStringParts();
-        foreach (array_keys($queryStringParts) as $key) {
-            if (!array_key_exists($key, $queryParams)) {
-                $queryParams[$key] = $queryStringParts[$key];
-            }
-        }
-        $queryParams[PropertyType::getCacheMissRelativeUrl()->getName()] = $this->_getPath($queryParams);
-        return $queryParams;
     }
 
     /**
@@ -68,18 +34,6 @@ class ConfigurationFetcher extends ConfigurationFetcherBase
             'languageVersion' =>
                 PHP_VERSION
         ]));
-    }
-
-    /**
-     * @param array $properties
-     * @return HttpResponseInterface
-     */
-    private function _fetchFromAPI(array $properties)
-    {
-        $url = $this->_getAPIUrl($properties);
-
-        $apiRequest = new RequestData($url, $properties);
-        return $this->_request->sendPost($apiRequest);
     }
 
     /**
