+++ src/Rox/Core/Network/ConfigurationFetcherOneSource.php
@@ -5,10 +5,6 @@ namespace Rox\Core\Network;
 use Exception;
 use Rox\Core\Client\BUIDInterface;
 use Rox\Core\Client\DevicePropertiesInterface;
-//use Rox\Core\Network\ConfigurationFetcherBase;
-//use Rox\Core\Network\ConfigurationSource;
-//use Rox\Core\Network\HttpClientInterface;
-//use Rox\Core\Network\HttpResponseInterface;
 use Rox\Core\Configuration\ConfigurationFetchedInvokerInterface;
 use Rox\Core\Reporting\ErrorReporterInterface;
 use Rox\Core\Consts\Environment;
@@ -52,12 +48,9 @@ class ConfigurationFetcherOneSource extends ConfigurationFetcherBase
     function fetch()
     {
         $usedSource = $this->getSource();
-
         try {
-            $queryParams = $this->_deviceProperties->getAllProperties();
-            $request = new RequestData($this->_environment->getConfigAPIPath(), $queryParams);
-            $fetchResult = $this->_request->sendGet($request);
-
+            $properties = $this->_preparePropsFromDeviceProps();
+            $fetchResult = $this->_fetchFromAPI($properties);
             if ($fetchResult->getStatusCode() == 200) {
                 return $this->createConfigurationResult($fetchResult->getContent()->readAsString(), $usedSource);
             } else {
